moving BI related service and controlle to the same namespace/folder to make it more self contained and usable as Microfronted
